// deuna-msa-nb-cnb-orq/src/domain/enums/document-validation.enum.ts
export enum DocumentValidationStatus {
  PENDING = 'PENDING',
  DONE = 'DONE',
  FAILED = 'FAILED',
}

export enum DocumentValidationResultStatus {
  APPROVED_VERIFIED = 'APPROVED_VERIFIED',
  DENIED_FRAUD = 'DENIED_FRAUD',
  DENIED_UNSUPPORTED_ID_TYPE = 'DENIED_UNSUPPORTED_ID_TYPE',
  DENIED_UNSUPPORTED_ID_COUNTRY = 'DENIED_UNSUPPORTED_ID_COUNTRY',
  ERROR_NOT_READABLE_ID = 'ERROR_NOT_READABLE_ID',
  NO_ID_UPLOADED = 'NO_ID_UPLOADED',
}

export enum statusResultValidationDocument {
  DONE = 'FINALIZED_OK',
  FAILED = 'FINALIZED_FAIL',
}